// src/app/shared/model/view.ts
export class View {
  nodeId?: any;
  supId?: string;
  title?: string;
  shortTitle?: string;
  description?: string;
  mediaFile?: any;
  supGuid?: string;
  bookMarkURL?: any;
  corPublicationStatus?: string;
  canAppearInGallery?: boolean;
  galleryDisplaySequence?: number;
  maxHLlevel?: number;
  defaultHLlevel?: number;
  vieIsGlobal?: boolean;
  spaceGuid?: any;
  tileFileName?: string;
  tileFilePath?: any;
  tileFileExtension?: string;
  viewDownloadLink?: string;
  datMediaExists?: boolean;
  precedence?: string;
  metViewCategory?: any;
  metViewGuid?: any;
  metViewTitle?: any;
  metViewShortTitle: any;
  queryTitle?: any;
  queryGuid?: any;
  viewMaskLabel?: string;
  queryConstructingView?: boolean;
}